Remove redundant BAM file open in paired mode
Opening a BAM file is an expensive operation as the index needs to be
fully read. In paired reads mode, at every contig change, the file was
being opened again to iterate over all reads from the previous contig.
This is usually not an issue for genome alignments, but transcriptome
alignments may have ~100k contigs, which makes this an expensive
operation.

Ideally, the two-pass mode should not have to read the file again, and
instead just maintain a rolling window of reads in memory.